backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sse-starlette==2.1.3
pydantic==2.9.2
redis[asyncio]==5.1.1
google-cloud-aiplatform==1.70.0
httpx==0.27.2
praw==7.7.1
pytrends==4.9.2
python-dotenv==1.0.1
